Refactor project structure for easier maintenance
- Move authentication-related functions to 'auth.py'.
- Keep password validation in 'security.py'.
- Remove hardcoded constants like SECRET_KEY, ALGORITHM, and ACCESS_TOKEN_EXPIRE_MINUTES and replace them with environment variables.
- Create specific routers for handling user functionality and authentication routes.
