+++ fast_todo/app.py
@@ -1,126 +1,13 @@
-from http import HTTPStatus
+from fastapi import FastAPI
 
-from fastapi import Depends, FastAPI, HTTPException
-
-# jwt
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy import select
-from sqlalchemy.orm import Session
-
-from fast_todo.database import get_session
-from fast_todo.models import User
-from fast_todo.schema import Message, Token, UserList, UserPublic, UserSchema
-from fast_todo.security import (
-    create_access_token,
-    get_current_user,
-    get_password_hash,
-    verify_password,
-)
+from fast_todo.routers import auth, users
+from fast_todo.schemas import Message
 
 app = FastAPI()
+app.include_router(users.router)
+app.include_router(auth.router)
 
 
 @app.get("/", response_model=Message)
 def read_root():
     return {"message": "Hello World!"}
-
-
-@app.post("/users/", status_code=HTTPStatus.CREATED, response_model=UserPublic)
-# Dependsにget_sessionを渡す（依存性の注入）
-def create_user(user: UserSchema, session: Session = Depends(get_session)):
-    db_user = session.scalar(
-        select(User).where(User.username == user.username)
-    )
-
-    if db_user:
-        raise HTTPException(
-            status_code=HTTPStatus.BAD_REQUEST,
-            detail="Username already exists",
-        )
-    hashed_password = get_password_hash(user.password)
-
-    db_user = User(
-        username=user.username, password=hashed_password, email=user.email
-    )
-    session.add(db_user)
-    session.commit()
-    session.refresh(db_user)
-
-    return db_user
-
-
-@app.get("/users/", response_model=UserList)
-def read_users(
-    skip: int = 0, limit: int = 100, session: Session = Depends(get_session)
-):
-    users = session.scalars(select(User).offset(skip).limit(limit)).all()
-    return {"users": users}
-
-
-@app.get("/users/{user_id}", response_model=UserPublic)
-def read_user(user_id: int, session: Session = Depends(get_session)):
-    db_user = session.scalar(select(User).where(User.id == user_id))
-    if not db_user:
-        raise HTTPException(
-            status_code=HTTPStatus.NOT_FOUND, detail="User not found"
-        )
-    return db_user
-
-
-@app.put("/users/{user_id}", response_model=UserPublic)
-def update_user(
-    user_id: int,
-    user: UserSchema,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    # 認可
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=HTTPStatus.BAD_REQUEST, detail="Not enough permissions"
-        )
-    current_user.username = user.username
-    current_user.email = user.email
-    current_user.password = get_password_hash(user.password)
-    session.commit()
-    session.refresh(current_user)
-
-    return current_user
-
-
-@app.delete("/users/{user_id}", response_model=Message)
-def delete_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=HTTPStatus.BAD_REQUEST, detail="Not enough permissions"
-        )
-    session.delete(current_user)
-    session.commit()
-    return {"message": "User deleted"}
-
-
-@app.post("/token", response_model=Token)
-def login_for_access_token(
-    form_data: OAuth2PasswordRequestForm = Depends(),
-    session: Session = Depends(get_session),
-):
-    user = session.scalar(select(User).where(User.email == form_data.username))
-
-    if not user:
-        raise HTTPException(
-            status_code=HTTPStatus.BAD_REQUEST,
-            detail="Incorrect email or password",
-        )
-
-    if not verify_password(form_data.password, user.password):
-        raise HTTPException(
-            status_code=HTTPStatus.BAD_REQUEST,
-            detail="Incorrect email or password",
-        )
-
-    access_token = create_access_token(data={"sub": user.username})
-    return {"access_token": access_token, "token_type": "bearer"}
