+++ tests/conftest.py
@@ -63,7 +63,7 @@ def user(session):
 @pytest.fixture()
 def token(client, user):
     response = client.post(
-        "/token",
+        "/auth/token",
         data={"username": user.email, "password": user.unhashed_password},
     )
     return response.json().get("access_token")
